@@ -1,7 +1,11 @@
-module.exports = {
-  content: ["./src/**/*.{js,ts,jsx,tsx}", "./index.html"],
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: [
+    "./src/**/*.{astro,html,md,mdx,js,ts,jsx,tsx}",
+    "./public/index.html",
+  ],
   theme: {
     extend: {},
   },
-  plugins: [],
+  plugins: [require("@tailwindcss/typography")],
 };
